split config file into multiple files
